FirstMapTest
If it imports into your project correctly, then go into the Tiled2Unity folder in Assests, then go to prefabs, and drag the map called Winter(something). Expand it in the hierarchy, and go to collisions. Expand collisions and turn of the mesh renderer. Then the thing right below this in the hierarchy is where the collision component is which can be modified.