Add files via upload
Drag out the map in the folder "prefabs". The redish looking tiles are collision tiles. In a folder called collisions, within the prefab, there are two game objects. One of them has the mesh renderer to show the redness of the collision. The other game object has the collision itself which can be fine tuned.